workflow-control.js: Change thisScriptObject to a string containing the function definition to avoid cross-context issues.
