+++ release-notes-5.0.0.md
@@ -14,4 +14,7 @@ to be included in customization scripts. Existing Nashorn scripts continue to be
 * Due to moving to Graal.js all `conditionFunction` elements in Workflows must be updated so that the function is named 
 `condition` and be Graal.js compatible.
 * Any `context.js` scripts must be Graal.js compatible.
+* The `thisScriptObject` field exposed by `workflow-control.js` is no longer an object. It is now a string containing 
+the `addFailures` function definition. Consumers must be updated to `eval()` this string into their context in order to 
+access the function. 
 * All downstream Workers must be upgraded to Worker-Document framework 4.5.0 to support executing Graal.js scripts.
